+++ engine/graphics/src/null/graphics_null.cpp
@@ -1839,7 +1839,7 @@ namespace dmGraphics
         return false;
     }
 
-    bool UnmapIndexBuffer(HContext context, HIndexBuffer buffer)
+    bool NullUnmapIndexBuffer(HContext context, HIndexBuffer buffer)
     {
         IndexBuffer* ib = (IndexBuffer*)buffer;
         memcpy(ib->m_Buffer, ib->m_Copy, ib->m_Size);
@@ -1848,7 +1848,7 @@ namespace dmGraphics
         return true;
     }
 
-    void* MapVertexBuffer(HContext context, HVertexBuffer buffer, BufferAccess access)
+    void* NullMapVertexBuffer(HContext context, HVertexBuffer buffer, BufferAccess access)
     {
         VertexBuffer* vb = (VertexBuffer*)buffer;
         vb->m_Copy = new char[vb->m_Size];
@@ -1856,7 +1856,7 @@ namespace dmGraphics
         return vb->m_Copy;
     }
 
-    bool UnmapVertexBuffer(HContext context, HVertexBuffer buffer)
+    bool NullUnmapVertexBuffer(HContext context, HVertexBuffer buffer)
     {
         VertexBuffer* vb = (VertexBuffer*)buffer;
         memcpy(vb->m_Buffer, vb->m_Copy, vb->m_Size);
@@ -1865,7 +1865,7 @@ namespace dmGraphics
         return true;
     }
 
-    void* MapIndexBuffer(HContext context, HIndexBuffer buffer, BufferAccess access)
+    void* NullMapIndexBuffer(HContext context, HIndexBuffer buffer, BufferAccess access)
     {
         IndexBuffer* ib = (IndexBuffer*)buffer;
         ib->m_Copy = new char[ib->m_Size];
